docs: generate_contrib_docs ext
